requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
python-dotenv
gunicorn
aiohttp
pydantic
starlette
redis
greenlet
itsdangerous
aiogram